requirements.txt
numpy==1.19.5
matplotlib==3.2.1
scikit-learn==1.0.2
Pillow==8.4.0